+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.1.0] - 2021-10-21
+
+### Added
+
+- AdSense wrapper class
+- Error handling for AdSense
+- Wortal prefab with auto-launcher
+
 ## [1.0.1] - 2021-10-14
 
 ### Added